MLProjects/CreditCardFraudDetection/requirement.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.0
python-multipart==0.0.7

# ML and analytics
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
xgboost==2.0.3
torch==2.1.2

# LLM and vector database
langchain==0.1.5
langchain-openai==0.0.5
langchain-community==0.0.16
pinecone-client==3.0.0
sentence-transformers==2.2.2
openai==1.12.0

# Monitoring and logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
httpx==0.26.0

# Production tools
gunicorn==21.2.0
redis==5.0.1
tenacity==8.2.3